requirements.txt
asgiref==3.5.2
bleach==5.0.1
Django==3.1.3
django-braces==1.15.0
django-model-utils==4.2.0
django-star-ratings==0.9.2
django-summernote==0.8.20.0
djangorestframework==3.13.1
Pillow==9.2.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
swapper==1.3.0
typing_extensions==4.3.0
webencodings==0.5.1
